[MOD] : 필요없는 코드 제거
