+++ app/src/main/java/com/sopt/now/compose/MainApplication.kt
@@ -1,0 +1,17 @@
+package com.sopt.now.compose
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+import timber.log.Timber
+
+@HiltAndroidApp
+class MainApplication : Application() {
+    override fun onCreate() {
+        super.onCreate()
+        setTimber()
+    }
+
+    private fun setTimber() {
+        Timber.plant(Timber.DebugTree())
+    }
+}
